Added new custom errors (AuthError, and ValError). All BL handlers are functional except for RegistrationHandler.

+++ src/logic/user/handlers/userhandler.ts
@@ -3,9 +3,12 @@ import { User } from "../../../data/user/user";
 import { Connection } from "typeorm";
 import { IServiceLocator } from "../../common/iservicelocator";
 import { UserRepository } from "../../../data/datamodule";
-import { TaskResult } from "../../common/results/taskresult";
+import { UserDeleteValidator } from "../../validation/user/validators/userdeletevalidator";
+import { ValidatorResult } from "../../validation/validatorresult";
+import { UserUpdateValidator } from "../../validation/user/validators/userupdatevalidator";
+import { ValidationError } from "../../validation/validationerror";
 
-/**
+/** 
  * Handler related to users. Manages users such as finding,
  * removing, or updating them.
  */
@@ -17,6 +20,16 @@ export class UserHandler extends LogicHandler {
     private userRepo: UserRepository;
 
     /**
+     * The validator to use when updating a user.
+     */
+    private updateValidator: UserUpdateValidator;
+
+    /**
+     * The validator to use when deleting a user.
+     */
+    private deleteValidator: UserDeleteValidator;
+
+    /**
      * Create a new user handler for managing users.
      * @param connection The database connection.
      * @param serviceLocator The dependency locator.
@@ -25,21 +38,21 @@ export class UserHandler extends LogicHandler {
         super(connection, serviceLocator);
 
         this.userRepo = connection.getCustomRepository(UserRepository);
+        this.updateValidator = new UserUpdateValidator();
+        this.deleteValidator = new UserDeleteValidator();
     }
 
     /**
-     * Checks if a username is available for taking.
+     * Checks if a username is available for taking.s
      * @param username The username to check for.
      * @returns True if the username is available.
      */
-    public async isUsernameAvailable(username: string):Promise<TaskResult<boolean>> {
+    public async isUsernameAvailable(username: string):Promise<boolean> {
         if(!username){
-            return TaskResult.errorResult('No username was passed in.');
+            throw new Error('No username was passed in.');
         }
-        else {
-            let isAvailable:boolean = await this.userRepo.isUsernameAvailable(username);
-            return TaskResult.successResult(isAvailable);
-        }
+        
+        return this.userRepo.isUsernameAvailable(username);
     }
 
     /**
@@ -48,14 +61,12 @@ export class UserHandler extends LogicHandler {
      * @param includeDeleted If we should include deleted users in the results.
      * @returns The user if found.
      */
-    public async findByUsername(username: string, includeDeleted?: boolean):Promise<TaskResult<User>> {
+    public async findByUsername(username: string, includeDeleted?: boolean):Promise<User> {
         if(!username){
-            return TaskResult.errorResult('No username was passed in');
+            throw new Error('No username was passed in');
         }
-        else {
-            let user: User = await this.userRepo.findByUsername(username, includeDeleted);
-            return TaskResult.successResult(user);
-        }
+
+        return this.userRepo.findByUsername(username, includeDeleted);
     }
 
     /**
@@ -65,14 +76,12 @@ export class UserHandler extends LogicHandler {
      * @param includeDeleted If we should include deleted users in the results.
      * @returns The user if found.
      */
-    public async findById(id: number, includeDeleted: boolean = false):Promise<TaskResult<User>> {
+    public async findById(id: number, includeDeleted: boolean = false):Promise<User> {
         if(isNaN(id)){
-            return TaskResult.errorResult('No user id passed in.');
+            throw new Error('No user id passed in.');
         }
-        else {
-            let user: User = await this.userRepo.findById(id, includeDeleted);
-            return TaskResult.successResult(user);
-        }
+        
+        return this.userRepo.findById(id, includeDeleted);
     }
 
     /**
@@ -81,11 +90,17 @@ export class UserHandler extends LogicHandler {
      * @returns True if no errors occured.
      */
     public async update(user: User): Promise<boolean> {
-        if(!user || isNaN(user.id) || user.isDeleted){
-            return false;
+        if(!user || isNaN(user.id)){
+            throw new Error('No user passed in, or user has no id.');
         }
         else {
-            return await this.userRepo.update(user);
+            let validatorResult: ValidatorResult = this.updateValidator.validate(user);
+
+            if(!validatorResult.isValid){
+                throw new ValidationError('Failed to update user.', validatorResult);
+            }
+            
+            return this.userRepo.update(user);
         }
     }
 
@@ -94,14 +109,18 @@ export class UserHandler extends LogicHandler {
      * @param user The user to delete
      * @returns True if no errors occured.
      */
-    public async delete(user: User): Promise<TaskResult<boolean>> {
-        //Bad data
+    public async delete(user: User): Promise<boolean> {
         if(!user || isNaN(user.id)){
-            return TaskResult.errorResult('No user passed in, or user has no id.');
+            throw new Error('No user passed in, or user has no id.');
         }
         else {
-            await this.userRepo.delete(user);
-            return TaskResult.successResult(true);
+            let validatorResult: ValidatorResult = this.deleteValidator.validate(user);
+
+            if(!validatorResult.isValid){
+                throw new ValidationError('Failed to delete user.', validatorResult);
+            }
+
+            return this.userRepo.delete(user);
         }
     }
 }
